requirements.txt
h5py==3.1.0
keras==2.6.0
keras_nightly==2.5.0.dev2021032900
numpy==1.19.3
Pillow==8.4.0
